requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cfdiclient==1.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
cssselect2==0.7.0
diff-match-patch==20230430
Django==5.0.7
django-bootstrap4==24.4
django-cfdi==2.215
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-import-export==4.1.1
django-phonenumber-field==8.0.0
django-rest==0.8.7
django-rest-swagger==2.2.0
django-simple-history==3.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
Flask==3.0.3
fonttools==4.54.1
html5lib==1.1
humanize==4.11.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
lxml==5.4.0
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.1.2
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.23.0
pydyf==0.11.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.1.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
tinycss2==1.3.0
typing_extensions==4.14.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
zopfli==0.2.3
